decompressor: implement read_across_frames=True for ZstdDecompressor.decompress()
This required us to change code to loop over the input. The code
pattern here is starting to converge with other APIs. I'm really
tempted to delete a bunch of code and implement some of these APIs
in terms of the io reader...

And since we're reading multiple frames, we also have to add logic
to track the total output size so we don't exceed the max output
size.